Move pixel selector functionality to the standalone file interact.py
